Fixed aspect ratio bug in GlfwUtil.h and properly propagated it to the pixel shader.
Summary: This really fixes the aspect ratio. It takes care of a long standing bug in GlfwUtil.h where the window width was not getting updated in the window class, which in turn is needed to compute the proper aspect ratio. Also the aspect ratio needed to be applied to only one coordinate (x in our case) vertex shader.  Last, I added a facial aspect ratio to make the face appear correct relative to the original video.

Reviewed By: aparrapo

Differential Revision: D18936393

fbshipit-source-id: 57208d5afa2b184eea46f5fb0ec4933f54af18d9