Use folly::DtoaMode for folly::json::serialization_opts
Summary:
Use `folly::DtoaMode` and `folly::DtoaFlags` override fields when using `folly::json::serialization_opts`.

This is a step in the multistep process to remove `double_conversion`. This step uses the shim types for json serialization. Another diff will remove `double_conversion` entirely.

double-conversion is removed because 1/ it increases android .apk binary sizes by ~100KB due to the inclusion of its libdouble-conversion.so (T194496993) and 2/ there is equivalent functionality in the standard library that eliminates the need for this library.

Reviewed By: Gownta

Differential Revision: D59781348

fbshipit-source-id: 78968524ee7805ba2c10947dfa3f6c00b12e0953