Removing reference to compromised S3 bucket
Summary:
T216227583 refers to a S3 Bucket Takeover Can Be Used To Spread Malware. This README file is from an archived project but it points to it. We can just delete the reference, but we still need to work with the original bucket owner to really solve the security issue

Created from CodeHub with https://fburl.com/edit-in-codehub

Reviewed By: bkcabral

Differential Revision: D74012281

fbshipit-source-id: 30ef7cf95cc6e50a485c7df82f56129e5f2fa4da